Fix condition

+++ database/seeders/UpdateAdministratorPermissions.php
@@ -15,7 +15,10 @@ class UpdateAdministratorPermissions extends Seeder
      */
     public function run()
     {
-        Permission::truncate();
+        if (app()->environment('local', 'staging', 'testing')) {
+            Permission::truncate();
+        }
+
         $this->call(PermissionsTableSeeder::class);
         $roles = Role::where('name', 'Administrator')->get();
